- Added Color Class which (for now) converts a vec4 into an int - Improved BatchRenderer so that he will create another draw call   when there ar too many sprites (up to 500K) - Added Transform2D class for Renderable2Ds

+++ NW_Gamesmith/src/maths/mat4.h
@@ -204,8 +204,8 @@ namespace gamesmith { namespace maths {
 				elements[8] * elements[1] * elements[6] -
 				elements[8] * elements[2] * elements[5];
 
-			double determinant = elements[0] * temp[0] + elements[1] * temp[4] + elements[2] * temp[8] + elements[3] * temp[12];
-			determinant = 1.0 / determinant;
+			float determinant = elements[0] * temp[0] + elements[1] * temp[4] + elements[2] * temp[8] + elements[3] * temp[12];
+			determinant = 1.0f / determinant;
 
 			for (int i = 0; i < 4 * 4; i++)
 				elements[i] = temp[i] * determinant;
